requirements.txt
# Pincer dependencies
#
# Install with: pip install -r requirements.txt
#
# ARM64 note: sqlite-vec may need to be built from source if no wheel is available.
# If 'pip install sqlite-vec' fails, ensure gcc and make are installed:
#   sudo apt install gcc make
# The agent will fall back to numpy cosine similarity if sqlite-vec is unavailable.

# Core
requests
python-dotenv

# Telegram gateway
python-telegram-bot>=20.0

# Task scheduler
schedule

# Vector memory (optional but recommended)
fastembed
sqlite-vec

# Voice message transcription (optional)
# Also requires: sudo apt install ffmpeg
faster-whisper

# Document ingestion
beautifulsoup4
pymupdf

# Web search
ddgs
